+++ packages/terra-clinical-action-header/examples/index-examples/ActionHeader-MaximizeClose.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-alert */
 import React from 'react';
-import ActionHeader from 'terra-clinical-action-header';
+import ActionHeader from '../../src/ActionHeader';
 
 const ActionHeaderExample = () => (
   <div>