+++ packages/terra-clinical-action-header/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* webdriver.io tests for `terra-clinical-action-header`
 
 2.0.0 - (February 19, 2018)
 ----------